Rename ajaxAlert to ajaxPattern to indicate broader use cases
