--- styleguide/source/_patterns/03-organisms/by-template/ajax-alerts.json
@@ -1,5 +1,0 @@
-{
-  "ajaxAlerts": {
-    "endpoint":"../../assets/data/emergency-alerts.json"
-  }
-}
