+++ styleguide/source/_patterns/03-organisms/by-template/ajax-pattern.json
@@ -1,0 +1,6 @@
+{
+  "ajaxPattern": {
+    "endpoint":"../../assets/data/emergency-alerts.json",
+    "renderPattern":"@organisms/by-template/emergency-alerts.twig"
+  }
+}
