--- styleguide/source/assets/scss/03-organisms/_ajax-alerts.scss
@@ -1,8 +1,0 @@
-.ma__ajax-alerts {
-  margin-top: 0;
-
-
-  &--no-alerts {
-   display: none;
-  }
-}
