+++ styleguide/source/assets/scss/03-organisms/_ajax-pattern.scss
@@ -1,0 +1,10 @@
+.ma__ajax-pattern {
+  margin: 0;
+  padding: 0;
+  width: 100%;
+
+
+  &--hidden {
+   display: none;
+  }
+}
