+++ styleguide/source/assets/scss/index.scss
@@ -108,7 +108,7 @@
 @import "02-molecules/wait-time";
 
 // Organisms - components
-@import "03-organisms/ajax-alerts";
+@import "03-organisms/ajax-pattern";
 @import "03-organisms/action-card";
 @import "03-organisms/action-details";
 @import "03-organisms/action-finder";
